# Reuse a PathCalculator to substantially speed up routing queries agai…