…nst the same CH.

```
shortcuts in bristol_east
                        time:   [1.2296 ms 1.2569 ms 1.2905 ms]
                        change: [-50.414% -49.842% -49.172%] (p = 0.00 < 0.05)
                        Performance has improved.
shortcuts in bristol_west
                        time:   [3.8934 ms 3.9038 ms 3.9153 ms]
                        change: [-48.807% -47.506% -46.338%] (p = 0.00 < 0.05)
                        Performance has improved.
shortcuts in strasbourg time:   [2.8561 ms 2.8767 ms 2.9002 ms]
                        change: [-85.964% -85.814% -85.657%] (p = 0.00 < 0.05)
                        Performance has improved.
```

The speedup is enormous -- 50% in Bristol, 85% in Strasbourg. It's roughly proportional to the size of the graph. The path calculator is some massive vector per node in the CH, so constantly allocating that doesn't make sense.

I didn't do this originally, just so we could have the satisfaction of getting the speedup now. ;)