@dabreegster 2025-02-14
> We may have to revisit if we end up using webworkers/multithreading for our routing computations.

Thread-local PathCalculators are the way to go then. There has to be a little care when using rayon to only initialize it once per thread, when the work-stealing stuff is enabled. But it's something I've figured out before: https://github.com/Urban-Analytics-Technology-Platform/od2net/blob/186c111f8cb5d006eaa33dfb03fc4faef9a65092/od2net/src/router.rs#L42

The hard part about multithreading will be getting rayon + vite happy, from vague memory...